prompted_responder/requirements.txt
llama-cpp-python==0.2.6
requests==2.31.0
